Sprint3 frontpage

+++ groupproject/views.py
@@ -1,5 +1,7 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render, redirect, get_object_or_404
+
+from accounts.mixins import StaffRequiredMixin
 from reviewertools.models import Review
 from .forms import PublisherForm, GameForm, DeveloperForm, GenreForm, PlatformForm, LanguageForm, ImageForm
 from django.views.generic import ListView, DetailView
@@ -9,8 +11,12 @@ from groupproject.models import Game, Publisher, Developer, Platform, Genre, Lan
 from django.urls import reverse_lazy
 from django.contrib import messages
 
-
 def DatabaseLinks(request):
+
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
+
     context = {
         'site_names': [('games', 'Game List'),
                        ('developers', 'Developer List'),
@@ -27,6 +33,9 @@ def DatabaseLinks(request):
 
 
 def CreatePublisher(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = PublisherForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -38,6 +47,9 @@ def CreatePublisher(request):
 
 
 def CreateDeveloper(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = DeveloperForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -49,6 +61,9 @@ def CreateDeveloper(request):
 
 
 def CreatePlatform(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = PlatformForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -60,6 +75,9 @@ def CreatePlatform(request):
 
 
 def CreateGenre(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = GenreForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -71,6 +89,9 @@ def CreateGenre(request):
 
 
 def CreateLanguage(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = LanguageForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -97,6 +118,9 @@ def ViewGenre(request, pk, template_name='data/genre/genre.html'):
     return render(request, template_name, context)
 
 def UpdateGenre(request, pk, template_name='data/genre/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     genre = get_object_or_404(Genre, pk=pk)
     context['genre'] = genre
@@ -126,6 +150,9 @@ def ViewLanguage(request, pk, template_name='data/language/language.html'):
 
 
 def UpdateLanguage(request, pk, template_name='data/language/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     language = get_object_or_404(Language, pk=pk)
     context['language'] = language
@@ -138,6 +165,9 @@ def UpdateLanguage(request, pk, template_name='data/language/update.html'):
     return render(request, template_name, context)
 
 def DeleteLanguage(request, pk, template_name='data/language/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     obj = get_object_or_404(Language, pk=pk)
     context['language'] = obj
@@ -148,6 +178,9 @@ def DeleteLanguage(request, pk, template_name='data/language/delete.html'):
     return render(request, template_name, context)
 
 def CreateGame(request):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     form = GameForm(request.POST or None, request.FILES or None)
     if form.is_valid():
@@ -170,6 +203,9 @@ def ViewGame(request, pk, template_name='data/game/game.html'):
     return render(request, template_name, context)
 
 def DeleteGenre(request, pk, template_name='data/genre/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     obj = get_object_or_404(Genre, pk=pk)
     context['genre'] = obj
@@ -180,6 +216,9 @@ def DeleteGenre(request, pk, template_name='data/genre/delete.html'):
     return render(request, template_name, context)
 
 def UpdateGame(request, pk, template_name='data/game/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     game = get_object_or_404(Game, pk=pk)
     context['game'] = game
@@ -193,6 +232,9 @@ def UpdateGame(request, pk, template_name='data/game/update.html'):
 
 
 def DeleteGame(request, pk, template_name='data/game/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     obj = get_object_or_404(Game, pk=pk)
     context['game'] = obj
@@ -225,21 +267,28 @@ def ViewPublisher(request, pk, template_name='data/publisher/publisher.html'):
 
 
 def UpdatePublisher(request, pk, template_name='data/publisher/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     publisher = get_object_or_404(Publisher, pk=pk)
     context['publisher'] = publisher
     form = PublisherForm(request.POST or None, instance=publisher)
-    context['form'] = form
+
     if form.is_valid():
         form.save()
         messages.success(request, 'Successfully Updated')
         return redirect('publisher_list')
+    context['form'] = form
     return render(request, template_name, context)
 
 
 def DeletePublisher(request, pk, template_name='data/publisher/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
-    obj = get_object_or_404(Publisher, pk)
+    obj = get_object_or_404(Publisher, pk=pk)
     context['publisher'] = obj
     if request.method == "POST":
         obj.delete()
@@ -256,19 +305,25 @@ def ViewDeveloper(request, pk, template_name='data/developer/developer.html'):
 
 
 def UpdateDeveloper(request, pk, template_name='data/developer/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
-    developer = get_object_or_404(Publisher, pk=pk)
+    developer = get_object_or_404(Developer, pk=pk)
     context['developer'] = developer
-    form = DeveloperForm(request.POST or None, instance=developer)
-    context['form'] = form
+    form = DeveloperForm(request.POST, instance=developer)
     if form.is_valid():
         form.save()
         messages.success(request, 'Successfully Updated')
         return redirect('developer_list')
+    context['form'] = form
     return render(request, template_name, context)
 
 
 def DeleteDeveloper(request, pk, template_name='data/developer/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     obj = get_object_or_404(Developer, pk=pk)
     context['developer'] = obj
@@ -294,6 +349,9 @@ def ViewPlatform(request, pk, template_name='data/platform/platform.html'):
 
 
 def UpdatePlatform(request, pk, template_name='data/platform/update.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     platform = get_object_or_404(Platform, pk=pk)
     context['developer'] = platform
@@ -307,6 +365,9 @@ def UpdatePlatform(request, pk, template_name='data/platform/update.html'):
 
 
 def DeletePlatform(request, pk, template_name='data/platform/delete.html'):
+    result = StaffRequiredMixin().dispatch(request)
+    if not result == None:
+        return result
     context = {}
     obj = get_object_or_404(Platform, pk=pk)
     context['developer'] = obj
